third_party/blink/web_tests/http/tests/preload/warning/sync-request.html

<!doctype html>
<html>
<link rel="preload" as="fetch" crossorigin="anonymous" href="resources/hello.txt"></link>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<body>
<script>
test(() => {
  const xhr = new XMLHttpRequest;
  xhr.open('GET', 'resources/hello.txt', false);
  xhr.send();
}, "Preload match fails because the request is synchronous (see the console log).");
</script>
</body>
</html>
